Define a Storage type
All the stuff that happens in settings.HOME_DIR should happen in the
Storage class.  This way we can move the I/O out of models.Build.  It
also separates the Django-specific parts from the main logic.  To
facilitate this we also create a Build dataclass that Storage works on
(and renamed models.Build to models.BuildModel).  Now all the storage
stuff can be manipulated/tested independent of the database.