Handle Build content "dynamically"
Instead of having a bunch of checks/methods for all the content in a
Build artifact, have a Build artifcat know what it's contents are and
handle them all the same way.  This assumes all build artifact contents
are to be handled the same way, so that has been made an acceptible
design decision.  It greatly simplifies the code.