tasks: Remove storage files as well as model on purge
Before the model's .delete() method was overridden so that it also
removed the files in storage, but this is no longer the case but the
purge tasks is only calling the model's .delete() method. It instead
needs to call BuildMan's .delete() method which knows how to delete the
database record and the storage.