tasks: pull_build: move most of the logic into BuildMan
At the time this task was created there was no BuildMan, so the tasks
took care of all the storage, db and jenkins. Now that BuildMan exists
it's pull() method should be doing this.  The one exception is the
task_id db field, which only the task knows about.

I may consider getting rid of the task_id field, as it's not really
being used for anything.