signals: delete files in an async task
If you delete a build from the Django admin, then the post-delete signal
for the django model starts deleting the build files from the tree.
This can take a long  time and cause the HTTP request to time out.

Instead have the signal kick off a (celery) task to delete the files.
Note this also means that the purge task will kick off delete tasks. I
think this should be ok.