managers: download, then extract build
When pulling a build from Jenkins, the request's `.iter_content`
iterator is "piped" to the storage's `extract_artifact` method, so the
request's connection is not closed until the iterator is completely
consumed which is when the download is fully extracted & rsynced.  On
some builds, on a busy system, this can take more than 7 minutes.
However Jenkins times out requests after 5(?) minutes of inactivity.

Make it so that we first download the entire build from Jenkins, then
pass the downloaded file to `extract_artifact`.  A wrapper around a
tempfile can be used such that the internal API doesn't have to change.

So now, e.g. we get
```
[2021-09-18 09:16:29,662: INFO/ForkPoolWorker-1] Pulling build: lighthouse.2928
[2021-09-18 09:16:29,662: INFO/ForkPoolWorker-1] Downloading build: lighthouse.2928
[2021-09-18 09:17:23,109: INFO/ForkPoolWorker-1] Downloaded build: lighthouse.2928
[2021-09-18 09:17:27,980: INFO/ForkPoolWorker-1] Extracting build: lighthouse.2928
[2021-09-18 09:23:33,704: INFO/ForkPoolWorker-1] Extracted build: lighthouse.2928
[2021-09-18 09:23:34,044: INFO/ForkPoolWorker-1] Pulled build lighthouse.2928
```
Where downloading and extracting are done in separate steps and so the
HTTP connection to Jenkins is closed much sooner.