managers: refactor "Build" into "BuildPublisher"
Instead of each individual "Build" being aware of its storage, Jenkins,
and database, have a "BuildPublisher" which knows about its storage,
Jenkins database and "builds".  This more reflects reality. And actually
the concept of a "build" goes away except for within the domain of the
BuildPublisher. You give BuildPublisher the BuildID and it puts
everything together to create the fantasy of a "build".