models: latest_build: compensate for legacy installs
We need to compensate for my legacy GBP instance where not all build
records have the `built` timestamp. That information is long gone and
can't be backported.

In such cases we rely on the number field for finding the latest build
if all builds for the given machine have no `built` timestamp. If any
build has a `built` timestamp then we only consider the ones with
`built` when calculating the latest build.