publisher: emit a signal when build is pulled
Emit a signal so that third parties can be informed when a build is
pulled.  This can, e.g. be used to send notifications.  Relevent build
information (packages, metadata) are sent to the event handler.  Use
python-dispatch instead of signals library included with Django so that
events registered with non-Django apps.