publisher: incomplete pull is not pulled
The Publisher has a pulled() method where it checks Storage to see if all
the components are there and returns True if they are.  However in the
dashboard there's a "Queue" list that checks to see if the Record's
.complete attribute is set.  In rare occurrence the Storage might be ok
but the Record is not.  For example the artifact got (possibly
partially) extracted but _update_build_metadata did not complete due to
server shutdown/crash.  In this case Publisher.pulled() should not
consider the build pulled.  Therefore check both the Storage status and
the Record field.

In theory we should only need to test the Record field since completed
implies pulled but again there may be a rare occurrence when this is not
the case.