tasks: pull: delete everything on Jenkins 404
Instead of just calling publisher.records.delete, which only deletes the
record, call publisher.delete which deletes the record and the storage.
This ensures that if there is no trace of the build post-deletion.

Since publisher.pull already checks that the build has been pulled
first, there's no danger of accidentally deleting a pulled build.