Embed celery config in the package distribution
I don't like how celery.py is part of the Django project and not the app
configuration.  Thoush Celery documentation prefers it that way, having
it part of the app makes it easier to install Gentoo Build Publisher,
and in deployments it's likely to be the only app with celery tasks
anyway.