builds: order machine builds by built date instead of submitted
Builds can get pulled in a different order than they were built.  For
example CICD could automatically pull build 10001 but afterwards one
might pull build 10000 for posterity.  However if that happens then `gbp
list` looks like this:

    $ gbp list lighthouse
    ...
    [*   ]  9987 08/31/22 22:23:38
    [*   ]  9998 09/01/22 09:12:41
    [*   ]  9999 09/01/22 10:10:24
    [* P ] 10001 09/01/22 11:12:09
    [    ] 10000 09/01/22 11:28:12

In the above example build 10001 was pulled (submitted) at 11:12 which
was done automatically by the Jenkins pipeline. However build 10000 was
pulled manually some time later, and that makes `gpg list` show it last.
It's gentoo_build_publisher that's actually doing the sorting. It is
sorting by the submitted timestamp.  The built timestamp was added
later.

This change has it build by the built timestamp.  Because there may
still be some old builds with no built timestamp, we first sort by built
then submitted ensuring that when submitted is null it sorts as older
and then all the nulls are sorted by the submitted timestamp.  The null
thing will probably only ever be a problem for me ;-)

RecordDB.latest() aready works because it does the right thing... sort
of. It checks if built is null and if so falls back to the build_id
however that may not work for non-Jenkins builds...