cli: introduce addrepo subcommand
Introduce "addrepo" as a gbpcli subcommand. This allows a user (on the
localhost) to add a repo job to the jenkins server, for example:

```
gbp addrepo gentoo https://anongit.gentoo.org/git/repo/gentoo.git
```

This is a lot easier than going through the Jenkins UI to do the same.