remove Django signals
Remove Django signals for post-delete as it only works with the Django
records backend.  The proper way to delete a Build is

BuildPublisher.delete(build)

This deletes the storage and the record (and anything else that needs
doing). Deleting a build any other way (e.g. Django admin) is not
recommended.  If things get out of sync then it will go detected by `gbp
check`.