publisher: don't use tempfile for jenkins download
Instead of downloading to a tempfile, then calling
Storage.extract_artifact which is taking the tempfile and writing it to
another tempfile, simply pass extract_artifact the download stream.

The reason it was done that way before was to work around a previous
issue which seems to have been solved (a missed flush() call,
perhaps?).