add ability to pull and add a note
At the Publisher and GraphQL layers, the pull function has an optional
new argument, `note`. When this argument is passed as a string, then the
BuildRecord is saved with the given string.

The reason for this is that sometimes I want to do a pull and add a note
to the build. Since pull is run async, a `pull` followed by a `note`
does not always work. Pulling with a note argument solves this issue.

Unanswered question: what, if anything, to do about `publish`?