cli: create the worker gbpcli subcommand
Gentoo Build Publisher now supports different task queue backends (3
included!). So in order to facility a central way to start the task
worker this commit introduces the "worker" gbpcli subcommand, so:

    $ gbp worker

When run the subcommand first determines which worker type is configured
(through Settings but can be overriden with a command-line argument). It
then calls the `work` classmethod on the appropriate worker interface.
This classmethod is responsible for instantiating the worker.

The "sync" backend doesn't have a worker so it exits with nonzero status
when invoked.