worker: allow workers to run any arbirary function
Instead of creating individual functions and registering them with the
worker, give the workers a single function, "run" and pass it the real
function to run.

For the "rq" and "sync" workers this is pretty much business as usual.
But for the Celery worker the function has to be marshalled and, for
security, signed, and since the Django signer only signs strings types
it also has to be base64-encoded before being sent to the Celery worker.
Then the celery task must unsign, base64-decode and then unmarshall the
parameter to get the function before it is called.