graphql: add a requires_apikey decorator
This decorator is to be used by GraphQL resolvers that require
authentication. The decorator checks that the HTTP request has a Basic
Auth header and that the header's name and secret matches an ApiKey
record. If it does then the record's last_used field is updated and the
decorated resolver is called and returned. If not then a GraphQL error
is raised.