[#1] conditionally require API keys
API keys shouldn't be strictly necessary. In most cases users will be
running all services on the same machine (per the Install Guide) and
said machine will be a home machine not exposed to the Internet. Create
a setting to disable API keys and use a decorator for mutations that
conditional require an API key.