settings: automatically add registered plugins to INSTALLED_APPS
If a package registers an entry_point in the
"gentoo_build_publisher.apps" group that references a string, then that
string is added to `INSTALLED_APPS` (if it's not already there). This
alleviates the need of the user to have to do this manually, thus making
installation/uninstallation of GBP apps easier.