set default (dummy) env variabls in __init__
Our test runner is using os.environ.setdefault to set some environment
variables that gentoo_build_publisher.publisher requires to default
values if they are not defined.  I'm finding that other projects that
depend on GBP need to do this too.  Instead of reqiring this to be
copied over into the other projects we can set the defaults in __init__
before instantiating the publisher.