check: make uncompleted builds with missing content warnings
Before they were skipped, but it should not be silent. I think the
reasoning behing skipping is a build that is not completely pulled may
not have all the content, so that seems normal. But on server crash it
will never complete and you're stuck with a bad build. So it should at
least give a warning.  Besides an not-completed build is pretty much an
exception if/when you're running check, wish is something that is
configured to run at service startup.