plugins: add beginings of the plugins infrastructur
"New-style" plugins will be a dict definition in one entry point instead
of a bunch of entrypoints. It will encompass the following

- Django app
- graphql
- menus
- Django urls (views)
- Description
- Homepage
- etc.

It's intended to be extensible and less noisy than the many entrypoints.
The apps string will be supported after the next release and dropped
before the release after that.