graphql: allow Package to get back to Build
This turned out to be a much bigger change than anticipated.

We want the Package GraphQL type to be able to get to the Build it
belongs to.

On the surface this seems pointless as the only way to get to the
Package is via the Build. However I gbp-fl would benefit from this as
its queries return package data more than build data.

So we add a "build" field to the Package dataclass. But in doing so many
functions need to have a BuildRecord passed to it.  But also the
metadata file, gbp.json shouldn't have the BuildRecord in the packages
built list (it's redundant). Also a lot of fucntions that accepted Build
or BuildRecord as an argument now only accept BuildRecord. This is
mostly not an issue as most of the time they are being called with
BuildRBuildRecord to begin with.