refactor: move django bits into a sub-package
Typically Django apps take the center stage, however I've never liked
this pattern. Gentoo Build Publisher is not a Django app. It's a system
of different components, one of which is Django. But the center of
Gentoo Build Publisher is not Django (the center is
.publisher.BuildPublisher, which knows nothing about Django).

With this change, the Django bits have been moved to a subdirectory
(sub-package). Models, views, etc are in the
django/gentoo_build_publisher directory. The reason for pushing it two
levels down instead of just one is to preserve the app (directory) name.
I didn't want a gentoo_build_publisher directory under
gentoo_build_publisher because I wanted "django" somewhere in the path
to denote that it's the Django bits.