deployment: config: add /home/gbp to PYTHONPATH
When __init__.py imported Celery, Celery adds the current working
directory to the Python path, which in the context of the systemd unit
files, allows Django to find the settings module.  However since Celery
imports were removed from __init__.py the CWD was not added and
subsequently Django fails to find the settings module.

This change adds /home/gbp to PYTHONPATH in gentoo-build-publisher.conf
so that Django can find the settings module again.