views: create_build_context: account for missing gbp.json
This is for very old builds before the introduction of gbp.json. The
build details page will 500 when attempting to get the packages built
(which is in the missing JSON file).  I think I only have one remaining
build that has this issue but am introducing this fix for it.