django: get rid of missing gbp.json errors once and for all
For really old builds there is no gbp.json file located in the
artifacts.  This causes views that depend on it (e.g. looking for
packages built) to fail. There have been try/excepts added to work
around this. However the GBPMetadata can be calculated on demand. So
this change adds a function to to just that. Instead of caching the
LookupError we now call the new function which always returns
GBPMetadata.

Note this is only for the Django views. The GraphQL later raised an
error, and still raises an error, when no gbp.json is found.  This may
or may not get fixed.