publisher.build_metadata: don't pull metadata from Jenkins
If gbp.json is missing and we need to generate GBPMetadata we can't
depend on the JenkinsMetadata being on the Jenkins server. This is
because the build may have been removed. Therefore this change,
calculates the JenkinsMetadata.duration (the only piece used in
GBPMetadata). It's an estimate but usually very close anyway.