plugins: make app optional
The "app" attribute on a Plugin (or PluginDef) tells the plugin system
which Django apps too add to INSTALLED_APPS. This was a required
attribute (could not be `None`). However some plugins, e.g.,
gbp-archive, may want to register as a plugin but have no Django app (or
urlpatterns, which is seperate) to expose. With this change these
plugins would be able to register themselves.