signals: switch from python-dispatch to blinker
With this change, the signals subsystem uses the blinker library. The
python-dispatch library hasn't had any activity for the past couple of
years. It also seems to have issues with the forthcoming Python 3.14.

This change replaces the python-dispatch API with an adapter so that,
e.g. plugins do not have to make any changes to switch to the blinker
library. The only caveat is that blinker does not support positional
arguments passed to signal handlers. Luckily Gentoo Build Publisher's
signals only send keyword arguments, so no changes were necessary.