signals.PyDispatcherAdapter: allow a handler to handler multiple events
Allow a single event handler to handle multiple events. This is a fix to
a bug found in gbp-fl[1] where subsequend calls to `.bind()` using the
same handler would allow the previously bound signals to get garbage
collected.

Also tests were not included when `PyDispatcherAdapter` was introduced
so this change adds needed tests.

[1] https://github.com/enku/gbp-fl/commit/565c0ae1d69572814eb145d21929a2e174bdb62e