utils.validate_identifier: don't validate invalid chars
The utils.validate_identifier function was testing that the given string
doesn't start with an invalid start character, however it was not
checking that the first character was any of the other invalid
characters. For example "@foo" checked as a valid identifier, allowing
"@" as the first character but not in the remaining characters when it
should not have been allowed at all.