cli.check: introduce checks as a plugin component
This change is so that plugins can implement their own checks that can
be run the the `gbp check` command.

The plugin definition adds a new field, `checks`, that when defined is a
dictionary of check_name => Check values. The CLI check subcommand then
iterates over each plugin, gets their checks, if any, and runs the
checks. Checks are specified in the PluginDef the same way as entry
points. In fact the subcommand uses `importlib.metadata.EntryPoint` to
load the checks. So in effect checks are basically entry points that are
not declared in the distribution.

Since Gentoo Build Publisher is a plugin of itself, it now defines its
own checks as plugin checks.