checks: move checks to their own module
Since the CLI subcommand to run the checks and the checks themselves are
separate components, it makes sense to have the checks live in their own
module. In fact this should be the convention for plugin checks: define
them in a `checks.py` module at the root of the package.