views: use straight ariadne instead of ariadne-django
We can create a Django view that converts the incoming request to a WSGI
request and pass that straight to the ariadne wsgi application. The
ariadne-django project has not been updated in three years and the
ariadne documentation recommends using their middleware and forking
Django requests to the Django WSGI app. We're doing that but the other
way around.