views: move StatsCollector into a seperate module
Move StatsCollector into a stats.py module. This is a step in the
eventual replacing of stats collecting from the views to a signal
handler, which will cache the stats for views to use.

As such we've also removed the cache argument from the StatsCollector.
There will be a seperate mechanism for caching the stats.