Package: ensure build field is not a BuildRecord
When Package.build is a BuildRecord then strange things happen, like the
entire BuildRecord being stored in gbp.json and the issue with
gbp-notifications' webhook method sending too large a payload in the
HTTP request. For this reason we restrict Package.build to only be Build
instance and raise TypeError when it is instantiated with a BuildRecord
instance.