cache: add the ability for subcaches to have a timeout
Add a set_timeout() function that, when given a subcache, set's a
timeout for the keys. E.g.

    mycache = cache / "mine"
    set_timeout(mycache, 300)
    mycache.foo = "bar"

In this case all the keys in mycache will expire after 30 seconds of
age. On the other hand the "root" cache
(gentoo_build_publisher.cache.cache) has no timeout.

This will help in migrating gbp-ps to site cache and getting rid of it's
various backends without the need to have a (direct) Django dependency.