graphql: add more stats
There commit ads a new stats Query field as well as adds a stats-related
field onto the MachineSummary type.

Note that totalPackageSize is a String instead of an Int because the
integer value will typically fall outside of the 32-bit signed int range
that GraphQL supports.  Callers will need to convert to their local
bigger integer type as necessary.