graphql: add tagInfo field to MachineSummary
The MachineSummary has a tags field that is a list of tags given to
builds for the given machine. However since it's a list of strings,
there is no way to tie the tag to the particular build without doing
another query. This change adds a new field, tagInfo, that returns a
list of type TagInfo that resolves both the tag string as well as the
Build object that the tag is attached to.