stats: don't cache MachineInfo
This change creates a MachineInfoDataclass class, which is nearly
identical to MachineInfo except the attributes are not (cached)
properties. Since MachineInfo's attributes are properties they are
evaluated lazily. So when acquiring the stats cache, some of the
attributes are not actually cached but evaluated as they are accessed.
So this means that the data you grab aren't really the "cached" data.
This also means that the "cached" stats depend on the database.  This
change uses a proxy class, MachineInfoDataclass, that greedily grabs
this data and caches it instead of the cached_property.