views: machine details: don't use the db
The machine details view (only) accesses the db to see if the requested
machine "exists" so it can raise Http404. With this change we change the
context helper so if the machine doesn't exist in the (cached) stats
then MachineNotFoundError is raised. The view can then catch this
exception and raise Http404.

The purpose is to enable the view to be accessible without the need of a
database hit.