tests: use fixtures.publisher instead of global one
The fixture returns the mocked BuildPublisher. Use that instead of
gentoo_build_publisher.publisher, even though that is patched as the
fixture. One of the advantages of using the fixture instead is that mypy
treats it's type as Any and so doesn't complain when you access
publisher.jenkins.artifact_builder.