web ui: allow user-configurable footer links
The links in the web ui filter were hard-coded and could not be
overriden by the user. With this change, GBP checks settings for the
FOOTER_LINKS name => URL link and uses that to render the links in the
footer. If that setting does not exist then a default is used.